EIP1-11740: Resolve PR comments
